movies_admin/requirements.txt
django==4.2.5
flake8==6.1.0
python-dotenv==1.0.0
psycopg==3.1.10
django_split_settings==1.2.0
djangorestframework==3.14.0
 drf-yasg==1.21.7
django-filter==23.3
django-cors-headers==4.2.0
